+++ libgtftk.h
@@ -325,6 +325,6 @@ void print_raw_data(RAW_DATA *raw_data, char delim);
 GTF_DATA *select_transcript(GTF_DATA *gtf_data, int type);
 SEQUENCES *get_sequences(GTF_DATA *gtf_data, char *genome_file, int intron, int rc);
 int free_gtf_data(GTF_DATA *gtf_data);
-void *get_memory(int size);
+void *get_memory(long int size);
 
 #endif /* GTFTOOLKIT_GTFTK_SRC_LIB_LIBGTFTK_H_ */
